Added Driver and SImpleSongBuilder
SimpleSongBuilder is the SimpleTableBuilder template from step 3. I changed the name since we're not building symbol tables, but rather songs. The driver has been updated to reflect this change and will not likely need more changes. The SimpleSongBuilder file is where most, if not all, of the work will be done.